Revert "splitting utils from tests and ensuring its excluded from coverage"
This reverts commit 7ce063e354758024ca8cca28423090c0ea0af63d.
